Build : upgrade parsing check file : .rt wtih out arg, too many arg check arg :one A C L check arg :at least one pl sp cy add : fov arg in cam

+++ srcs/parsing_utils.c
@@ -116,17 +116,17 @@ int	str_to_vecdir(char **line, t_vec4 *v)
 	char	*end;
 
 	str = *line;
-	v->x = (int)ft_strtof(str, &end);
+	v->x = ft_strtof(str, &end);
 	if (*end != ',' || str == end)
 		return (-2);
 	norm_float(&v->x, -1.0f, 1.0f);
 	str = end + 1;
-	v->y = (int)ft_strtof(str, &end);
+	v->y = ft_strtof(str, &end);
 	if (*end != ',' || str == end)
 		return (-2);
 	norm_float(&v->y, -1.0f, 1.0f);
 	str = end + 1;
-	v->z = (int)ft_strtof(str, &end);
+	v->z = ft_strtof(str, &end);
 	if (str == end)
 		return (-2);
 	norm_float(&v->z, -1.0f, 1.0f);
